Uppgift 3 Implementerat #region (fan vad fint).
