+++ Uppgift3/Klasser/Program.cs
@@ -9,45 +9,41 @@ namespace Klasser
 
         static void Main(string[] args)
         {
-            // Skapar fyra objekt av klassen Bil
-            Car BMW = new Car("BMW Z3")
-            {
-                WeightInKG = 1000,
-                Registered = DateTime.Now.ToString("yyyy/MM/dd"),
-                Electric = false,
-                LicensePlate = "ABC123"
-            };
-            Car Tesla = new Car("Tesla s-series")
-            {
-                WeightInKG = 920,
-                Registered = DateTime.Now.ToString("yyyy/MM/dd"),
-                Electric = true,
-                LicensePlate = "DEF456"
-            };
-            Car Volkswagen = new Car("Volskwagen Passat")
-            {
-                WeightInKG = 1100,
-                Registered = DateTime.Now.ToString("yyyy/MM/dd"),
-                Electric = false,
-                LicensePlate = "GHI789"
-            };
-            Car Chevrolet = new Car("Chevrolet Volt")
-            {
-                WeightInKG = 1150,
-                Registered = DateTime.Now.ToString("yyyy/MM/dd"),
-                Electric = true,
-                LicensePlate = "JKL101"
-            };
 
-            // Lista för de skapade objekten
+            #region Skapar fyra objekt av klassen Bil
+            Car BMW = new Car("BMW Z3",
+                1000,
+                DateTime.Now.ToString("yyyy/MM/dd"),
+                "ABC123",
+                false);
+            Car Tesla = new Car("Tesla s-series",
+                920,
+                DateTime.Now.ToString("yyyy/MM/dd"),
+                "DEF456",
+                true);
+            Car Volkswagen = new Car("Volskwagen Passat",
+                1100,
+                DateTime.Now.ToString("yyyy/MM/dd"),
+                "GHI789",
+                false);
+            Car Chevrolet = new Car("Chevrolet Volt",
+                1150,
+                DateTime.Now.ToString("yyyy/MM/dd"),
+                "JKL101",
+                true);
+            #endregion
+
+            // Skapar lista för de skapade bil-objekten
             List<Car> listOfCars = new List<Car>();
-            // Lägger till de skapade objekten i listan
+            #region Lägger till de skapade objekten i listan
             listOfCars.Add(BMW);
             listOfCars.Add(Tesla);
             listOfCars.Add(Volkswagen);
             listOfCars.Add(Chevrolet);
+            #endregion
 
-            // Skriver ut de skapade objekten
+
+            #region Skriver ut de skapade objekten
             Console.WriteLine("Skapade objekt innan programmet körs: ");
             PrintCars(listOfCars);
 
@@ -57,39 +53,42 @@ namespace Klasser
 
             // Tar bort alla element i listan
             listOfCars.Clear();
+            #endregion
 
             // Lista som tar emot objekt av klassen Person
             List<Person> listOfUsers = new List<Person>();
-            
+
             // Start av huvudprogram (skapa obegränsat antal personer och tilldela bilar till dem)
             var isAddingPersons = true;
             do
             {
+                #region Samlar in data om användaren
                 Console.WriteLine("\n\t----- LÄGG TILL PERSON -----");
                 var userName = SetName("Mata in namn: ");
                 var userAge = SetAgeOrWeight("Mata in ålder: ");
 
-                // Skapar objekt av klassen Person
-                Person user = new Person(userName, userAge);
-                // Lägger till user i lista med användare.
-                listOfUsers.Add(user);
+                Person user = new Person(userName, userAge); // Skapar objekt av klassen Person
+                listOfUsers.Add(user); // Lägger till user i lista med användare.
+                #endregion
 
                 var isAddingCars = true;
                 do
                 {
+                    #region Samlar in data om bilen
                     Console.WriteLine($"\n\t----- REGISTRERA BIL TILL ANVÄNDARE \"{user.GetName().ToUpper()}\" -----");
                     var modelName = SetName("Mata in modell: ");
                     var weight = SetAgeOrWeight("Mata in vikt: ");
                     var licensePlate = SetLicensePlate("Mata in registreringsnummer: ");
                     var isElectric = SetBool("Är det en elbil? j/n: ");
                     var registered = DateTime.Now.ToString("yyyy/MM/dd"); // Registrering av fordon sker när applikationen körs
+                    #endregion
 
                     // Skapar nytt objekt av klassen car.
                     Car car = new Car(modelName, weight, registered, licensePlate, isElectric);
+                    user.Cars.Add(car); // Lägger in car i användarlistan Cars
 
-                    // Lägger in car i användarlistan Cars
-                    user.Cars.Add(car);
-
+                    // Om användaren inte vill registrera fler bilar så frågar vi om den vill lägga till fler personer
+                    // Om inte, skriv ut all data vi har och avsluta.
                     Console.Write($"Vill du registrera fler bilar till {user.GetName()}? j/n: ");
                     var answer = Console.ReadLine().ToLower();
                     if (answer != "j")
@@ -106,16 +105,15 @@ namespace Klasser
                     }
 
                 } while (isAddingCars);
+
             } while (isAddingPersons);
-            
+
             PrintUsersAndCars(listOfUsers);
             Console.ReadKey(); // Avslutar program
         }
 
-        
+        #region Metoder
 
-
-        // Metoder
         /// <summary>
         /// Skriver ut alla bilar som finns i en lista.
         /// </summary>
@@ -180,10 +178,10 @@ namespace Klasser
                 Console.Write(question);
                 if (!int.TryParse(Console.ReadLine(), out ageOrWeight))
                     Console.WriteLine("Du måste mata in ett heltal!");
-                
+
                 else
                     isInputting = false;
-                
+
 
             } while (isInputting);
 
@@ -220,127 +218,15 @@ namespace Klasser
         /// <returns>Boolean</returns>
         public static bool SetBool(string questionJorN)
         {
-            
+
             Console.Write(questionJorN);
             var answer = Console.ReadLine().ToLower();
-            
+
             // Om svar är j/J så sätts electric till true.
             var electric = (answer == "j");
             return electric;
         }
-
-    }
-
-    public class Car
-    {
-        // Fields
-        public string _model;
-        private decimal _odometer;
-
-        // Properties
-        public int WeightInKG { get; set; }
-        public string Registered { get; set; }
-        public bool Electric { get; set; }
-        public string LicensePlate { get; set; }
-
-        /// <summary>
-        /// Adderar hur långt ett fordon har kört till milmätaren.
-        /// </summary>
-        /// <param name="lengthDriven">Hur långt fordonet har kört</param>
-        public void SetOdometer(decimal lengthDriven)
-        {
-            if (lengthDriven > 0)
-            {
-                _odometer += lengthDriven;
-            }
-            
-        }
-
-        /// <summary>
-        /// Hämtar milmätaren och omvandlar till <b>string</b>.
-        /// </summary>
-        /// <returns>String</returns>
-        public string GetOdometer()
-        {
-            return _odometer.ToString();
-        }
-
-        /// <summary>
-        /// Skriver ut all information om en bil.
-        /// </summary>
-        public void GetInfo()
-        {
-            Console.WriteLine();
-            Console.WriteLine($"Modell: {_model}");
-            Console.WriteLine($"Vikt: {WeightInKG}kg");
-            Console.WriteLine($"Registrerades: {Registered}");
-            Console.WriteLine($"Registreringsnummer: {LicensePlate}");
-            if (Electric)
-                Console.WriteLine("\"Detta är en elbil!\"");
-        }
-
-        /// <summary>
-        /// Sätter modellen på bilen.
-        /// </summary>
-        /// <param name="modelname">Modellnamn.</param>
-        public Car(string modelname)
-        {
-            _model = modelname;
-        }
-
-        /// <summary>
-        /// Sätter alla uppgifter om en bil.
-        /// </summary>
-        /// <param name="model"></param>
-        /// <param name="weight"></param>
-        /// <param name="registered"></param>
-        /// <param name="licenseplate"></param>
-        /// <param name="electric"></param>
-        public Car(string model, int weight,
-            string registered,
-            string licenseplate,
-            bool electric)
-        {
-            this._model = model;
-            this.WeightInKG = weight;
-            this.Registered = registered;
-            this.LicensePlate = licenseplate;
-            this.Electric = electric;
-        }
-
-    }
-
-    public class Person
-    {
-        private string name;
-        private int age;
-        public List<Car> Cars { get; set; }
-
-        /// <summary>
-        /// Hämtar en persons namn
-        /// </summary>
-        /// <returns>String</returns>
-        public string GetName()
-        {
-            return name;
-        }
-        /// <summary>
-        /// Hämtar en persons ålder
-        /// </summary>
-        /// <returns>Integer</returns>
-        public int GetAge()
-        {
-            return age;
-        }
-
-        
-        public Person(string name, int age)
-        {
-            this.name = name;
-            this.age = age;
-            Cars = new List<Car>();
-        }
-
+        #endregion
     }
 
 }
